refactor(array-utils): updated imports & packages to only use array-utils where applicable
