+++ packages/core/package.json
@@ -33,6 +33,7 @@
     "@jscad/modeling": "0.0.1",
     "@jscad/io": "1.0.0-alpha.d0fb3056",
     "@jscad/io-utils": "1.0.0-alpha.d0fb3056",
+    "@jscad/array-utils":"0.0.0",
     "@jscad/vtree":"0.0.1",
     "@most/create": "^2.0.1",
     "es2015-i18n-tag": "^1.3.1",
