+++ packages/io/io/package.json
@@ -53,6 +53,7 @@
     "@jscad/svg-deserializer": "1.0.0-alpha.d0fb3056",
     "@jscad/svg-serializer": "1.0.0-alpha.d0fb3056",
     "@jscad/x3d-serializer": "1.0.0-alpha.d0fb3056",
-    "@jscad/modeling": "0.0.1"
+    "@jscad/modeling": "0.0.1",
+    "@jscad/array-utils":"0.0.0"
   }
 }
