+++ packages/utils/regl-renderer/package.json
@@ -13,7 +13,8 @@
   "devDependencies": {
     "budo": "10.0.4",
     "@jscad/img-utils": "0.0.0",
-    "@jscad/scad-api":"latest"
+    "@jscad/scad-api":"latest",
+    "@jscad/array-utils":"0.0.0"
   },
   "dependencies": {
     "angle-normals": "1.0.0",
