+++ packages/web/package.json
@@ -44,6 +44,7 @@
     "@jscad/io": "1.0.0-alpha.d0fb3056",
     "@jscad/regl-renderer": "0.0.0",
     "@jscad/vtree":"0.0.1",
+    "@jscad/array-utils":"0.0.0",
     "bel": "^6.0.0",
     "brace": "0.11.1",
     "codemirror": "^5.37.0",
